requirements.txt
django == 3.2.16
django-debug-toolbar == 3.7.0 
mysqlclient == 2.1.1 
graphene-django == 2.15.0
django-filter == 22.1 
django-graphql-auth ==  0.3.16 
django-graphql-jwt ==  0.3.0
django-cors-headers == 3.13.0 
django-redis == 5.2.0

